+++ generatorSqlmapCustom/src/cn/exrick/manager/pojo/TbMemberExample.java
@@ -1,5 +1,6 @@
 package cn.exrick.manager.pojo;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
@@ -975,52 +976,52 @@ public class TbMemberExample {
             return (Criteria) this;
         }
 
-        public Criteria andBalanceEqualTo(Long value) {
+        public Criteria andBalanceEqualTo(BigDecimal value) {
             addCriterion("balance =", value, "balance");
             return (Criteria) this;
         }
 
-        public Criteria andBalanceNotEqualTo(Long value) {
+        public Criteria andBalanceNotEqualTo(BigDecimal value) {
             addCriterion("balance <>", value, "balance");
             return (Criteria) this;
         }
 
-        public Criteria andBalanceGreaterThan(Long value) {
+        public Criteria andBalanceGreaterThan(BigDecimal value) {
             addCriterion("balance >", value, "balance");
             return (Criteria) this;
         }
 
-        public Criteria andBalanceGreaterThanOrEqualTo(Long value) {
+        public Criteria andBalanceGreaterThanOrEqualTo(BigDecimal value) {
             addCriterion("balance >=", value, "balance");
             return (Criteria) this;
         }
 
-        public Criteria andBalanceLessThan(Long value) {
+        public Criteria andBalanceLessThan(BigDecimal value) {
             addCriterion("balance <", value, "balance");
             return (Criteria) this;
         }
 
-        public Criteria andBalanceLessThanOrEqualTo(Long value) {
+        public Criteria andBalanceLessThanOrEqualTo(BigDecimal value) {
             addCriterion("balance <=", value, "balance");
             return (Criteria) this;
         }
 
-        public Criteria andBalanceIn(List<Long> values) {
+        public Criteria andBalanceIn(List<BigDecimal> values) {
             addCriterion("balance in", values, "balance");
             return (Criteria) this;
         }
 
-        public Criteria andBalanceNotIn(List<Long> values) {
+        public Criteria andBalanceNotIn(List<BigDecimal> values) {
             addCriterion("balance not in", values, "balance");
             return (Criteria) this;
         }
 
-        public Criteria andBalanceBetween(Long value1, Long value2) {
+        public Criteria andBalanceBetween(BigDecimal value1, BigDecimal value2) {
             addCriterion("balance between", value1, value2, "balance");
             return (Criteria) this;
         }
 
-        public Criteria andBalanceNotBetween(Long value1, Long value2) {
+        public Criteria andBalanceNotBetween(BigDecimal value1, BigDecimal value2) {
             addCriterion("balance not between", value1, value2, "balance");
             return (Criteria) this;
         }
