+++ xmall-manager/xmall-manager-pojo/src/main/java/cn/exrick/manager/dto/front/CartProduct.java
@@ -1,12 +1,16 @@
 package cn.exrick.manager.dto.front;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
+/**
+ * @author Exrickx
+ */
 public class CartProduct implements Serializable{
 
     private Long productId;
 
-    private Long salePrice;
+    private BigDecimal salePrice;
 
     private Long productNum;
 
@@ -34,11 +38,11 @@ public class CartProduct implements Serializable{
         this.productId = productId;
     }
 
-    public Long getSalePrice() {
+    public BigDecimal getSalePrice() {
         return salePrice;
     }
 
-    public void setSalePrice(Long salePrice) {
+    public void setSalePrice(BigDecimal salePrice) {
         this.salePrice = salePrice;
     }
